code review: renaming funcs and packages
