+++ sensor/kubernetes/main.go
@@ -12,7 +12,7 @@ import (
 	"github.com/stackrox/rox/pkg/premain"
 	"github.com/stackrox/rox/pkg/utils"
 	"github.com/stackrox/rox/pkg/version"
-	"github.com/stackrox/rox/sensor/common/connection"
+	"github.com/stackrox/rox/sensor/common/centralclient"
 	"github.com/stackrox/rox/sensor/kubernetes/client"
 	"github.com/stackrox/rox/sensor/kubernetes/fake"
 	"github.com/stackrox/rox/sensor/kubernetes/sensor"
@@ -46,12 +46,12 @@ func main() {
 	} else {
 		sharedClientInterface = client.MustCreateInterface()
 	}
-	connFactory, err := connection.NewConnectionFactory(env.CentralEndpoint.Setting())
+	centralConnFactory, err := centralclient.NewCentralConnectionFactory(env.CentralEndpoint.Setting())
 	if err != nil {
 		utils.CrashOnError(errors.Wrapf(err, "sensor failed to start while initializing gRPC client to endpoint %s", env.CentralEndpoint.Setting()))
 	}
 
-	s, err := sensor.CreateSensor(sharedClientInterface, workloadManager, connFactory)
+	s, err := sensor.CreateSensor(sharedClientInterface, workloadManager, centralConnFactory)
 	utils.CrashOnError(err)
 
 	s.Start()
